Refactoring Handler Location
