--- Api/Core/Opss.PrimaryAuthorityRegister.Api.Application/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace Opss.PrimaryAuthorityRegister.Api.Application
-{
-    public class Class1
-    {
-
-    }
-}
