+++ Api/Core/Opss.PrimaryAuthorityRegister.Api.Application/Opss.PrimaryAuthorityRegister.Api.Application.csproj
@@ -6,4 +6,12 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Services\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\Common\Opss.PrimaryAuthorityRegister.Common\Opss.PrimaryAuthorityRegister.Common.csproj" />
+  </ItemGroup>
+
 </Project>
