--- Api/Presentation/Opss.PrimaryAuthorityRegister.Api/Handlers/TestHandlers/UpdateTestDataCommandHandler.cs
@@ -1,20 +1,0 @@
-﻿using MediatR;
-using Opss.PrimaryAuthorityRegister.Common.Requests.Test.Commands;
-
-namespace Opss.PrimaryAuthorityRegister.Api.Handlers.TestHandlers
-{
-    public class UpdateTestDataCommandHandler : IRequestHandler<UpdateTestDataCommand>
-    {
-        public UpdateTestDataCommandHandler(ILogger<UpdateTestDataCommandHandler> logger)
-        {
-        }
-
-        public async Task Handle(UpdateTestDataCommand request, CancellationToken cancellationToken)
-        {
-            ArgumentNullException.ThrowIfNull(request);
-
-            // do nothing
-            await Task.CompletedTask.ConfigureAwait(false);
-        }
-    }
-}
